feat(files) divide game.html into files
