+++ src/main/resources/static/js/game/stateManager.js
@@ -1,0 +1,9 @@
+export const state = {
+    roomId: null,
+    playerId: null,
+    currentPlayerId: null,
+    isSente: true,
+    socket: null,
+    board: [],
+    selectedFrom: null
+};
