+++ src/main/resources/templates/game.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title th:text="#{header.game_title}">将棋対局</title>
     <link rel="stylesheet" href="/css/game.css"/>
-    <script src="/js/game.js" defer></script>
+    <script type="module" src="/js/game/game.js" defer></script>
 </head>
 
 <body>
